Moved GenerateCreateTableStatement to the sqlfmt package
