+++ go/libraries/doltcore/diff/table_deltas.go
@@ -612,70 +612,6 @@ func GetDataDiffStatement(tableName string, sch schema.Schema, row sql.Row, rowD
 	}
 }
 
-// GenerateCreateTableStatement returns CREATE TABLE statement for given table. This function was made to share the same
-// 'create table' statement logic as GMS. We initially were running `SHOW CREATE TABLE` query to get the statement;
-// however, it cannot be done for cases that need this statement in sql shell mode. Dolt uses its own Schema and
-// Column and other object types which are not directly compatible with GMS, so we try to use as much shared logic
-// as possible with GMS to get 'create table' statement in Dolt.
-func GenerateCreateTableStatement(tblName string, sch schema.Schema, pkSchema sql.PrimaryKeySchema, fks []doltdb.ForeignKey, fksParentSch map[string]schema.Schema) (string, error) {
-	sqlSch := pkSchema.Schema
-	colStmts := make([]string, len(sqlSch))
-
-	// Statement creation parts for each column
-	for i, col := range sch.GetAllCols().GetColumns() {
-		colStmts[i] = sqlfmt.GenerateCreateTableIndentedColumnDefinition(col, sql.CollationID(sch.GetCollation()))
-	}
-
-	primaryKeyCols := sch.GetPKCols().GetColumnNames()
-	if len(primaryKeyCols) > 0 {
-		primaryKey := sql.GenerateCreateTablePrimaryKeyDefinition(primaryKeyCols)
-		colStmts = append(colStmts, primaryKey)
-	}
-
-	indexes := sch.Indexes().AllIndexes()
-	for _, index := range indexes {
-		// The primary key may or may not be declared as an index by the table. Don't print it twice if it's here.
-		if isPrimaryKeyIndex(index, sch) {
-			continue
-		}
-		colStmts = append(colStmts, sqlfmt.GenerateCreateTableIndexDefinition(index))
-	}
-
-	for _, fk := range fks {
-		colStmts = append(colStmts, sqlfmt.GenerateCreateTableForeignKeyDefinition(fk, sch, fksParentSch[fk.ReferencedTableName]))
-	}
-
-	for _, check := range sch.Checks().AllChecks() {
-		colStmts = append(colStmts, sqlfmt.GenerateCreateTableCheckConstraintClause(check))
-	}
-
-	coll := sql.CollationID(sch.GetCollation())
-	createTableStmt := sql.GenerateCreateTableStatement(tblName, colStmts, coll.CharacterSet().Name(), coll.Name())
-	return fmt.Sprintf("%s;", createTableStmt), nil
-}
-
-// isPrimaryKeyIndex returns whether the index given matches the table's primary key columns. Order is not considered.
-func isPrimaryKeyIndex(index schema.Index, sch schema.Schema) bool {
-	var pks = sch.GetPKCols().GetColumns()
-	var pkMap = make(map[string]struct{})
-	for _, c := range pks {
-		pkMap[c.Name] = struct{}{}
-	}
-
-	indexCols := index.ColumnNames()
-	if len(indexCols) != len(pks) {
-		return false
-	}
-
-	for _, c := range index.ColumnNames() {
-		if _, ok := pkMap[c]; !ok {
-			return false
-		}
-	}
-
-	return true
-}
-
 // WorkingSetContainsOnlyIgnoredTables returns true if all changes in working set are ignored tables.
 // Otherwise, if there are any non-ignored changes, returns false.
 // Note that only unstaged tables are subject to dolt_ignore (this is consistent with what git does.)
