+++ go/libraries/doltcore/sqle/enginetest/dolt_engine_test.go
@@ -204,35 +204,29 @@ func newUpdateResult(matched, updated int) gmstypes.OkResult {
 
 // Convenience test for debugging a single query. Unskip and set to the desired query.
 func TestSingleMergeScript(t *testing.T) {
-	t.Skip()
+	// t.Skip()
 	var scripts = []MergeScriptTest{
 		{
-			Name: "adding a non-null column with a default value to one side",
+			Name: "right-side adds a column with a default value",
 			AncSetUpScript: []string{
-				"set dolt_force_transaction_commit = on;",
-				"create table t (pk int primary key, col1 int);",
-				"insert into t values (1, 1);",
+				"CREATE table t (pk int primary key, c1 varchar(100), c2 varchar(100));",
+				"INSERT into t values ('1', 'BAD', 'hello');",
 			},
 			RightSetUpScript: []string{
-				"alter table t add column col2 int not null default 0",
-				"alter table t add column col3 int;",
-				"insert into t values (2, 2, 2, null);",
+				"alter table t add column c3 varchar(100) default (CONCAT(c2, c1, 'default'));",
+				"insert into t values ('2', 'BAD', 'hello', 'hi');",
 			},
 			LeftSetUpScript: []string{
-				"insert into t values (3, 3);",
+				"insert into t values ('3', 'BAD', 'hi');",
 			},
 			Assertions: []queries.ScriptTestAssertion{
 				{
 					Query:    "call dolt_merge('right');",
-					Expected: []sql.Row{{0, 0}},
+					Expected: []sql.Row{{doltCommit, 0, 0}},
 				},
 				{
-					Query:    "select * from t;",
-					Expected: []sql.Row{{1, 1, 0, nil}, {2, 2, 2, nil}, {3, 3, 0, nil}},
-				},
-				{
-					Query:    "select pk, violation_type from dolt_constraint_violations_t",
-					Expected: []sql.Row{},
+					Query:    "select * from t order by pk;",
+					Expected: []sql.Row{{1, "BAD", "hello", "helloBADdefault"}, {2, "BAD", "hello", "hi"}, {3, "BAD", "hi", "hiBADdefault"}},
 				},
 			},
 		},
